sierra_adapters reorganized to platform, solr, and comms modules
