+++ babel/sierra_adapters/solr.py
@@ -1,0 +1,10 @@
+"""
+This module handles requests to BPL Solr.
+"""
+
+from bookops_bpl_solr import SolrSession
+
+
+class BplSolr(SolrSession):
+    def __init__(self):
+        pass
